+++ api/src/modelos/usuario.go
@@ -11,7 +11,7 @@ import (
 
 // Usuario representa um usuário utilizando a rede social
 type Usuario struct {
-	ID       int       `json:"id,omitempty"`
+	ID       uint64    `json:"id,omitempty"`
 	Nome     string    `json:"nome,omitempty"`
 	Nick     string    `json:"nick,omitempty"`
 	Email    string    `json:"email,omitempty"`
